+++ Lenta/Models/SendPost.swift
@@ -8,7 +8,7 @@
 import UIKit
 
 struct SendPost {
-    let userId: Int
+    let userId: Int16
     let description: String
     let image: UIImage?
 }
